requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4            # via black, virtualenv
argon2-cffi==20.1.0       # via notebook
astroid==2.4.2            # via pylint
attrs==19.3.0             # via black, jsonschema, pytest
backcall==0.2.0           # via ipython
bandit==1.6.2             # via -r requirements-dev.in
black==19.10b0            # via -r requirements-dev.in
bleach==3.1.5             # via nbconvert
boto3==1.14.38            # via -r requirements-dev.in
botocore==1.17.38         # via boto3, s3transfer
certifi==2020.6.20        # via -c requirements.txt, requests, sentry-sdk
cffi==1.14.1              # via argon2-cffi
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via -c requirements.txt, requests
click==7.1.2              # via black, safety, wandb
configparser==5.0.0       # via wandb
cycler==0.10.0            # via -c requirements.txt, matplotlib
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
distlib==0.3.1            # via virtualenv
docker-pycreds==0.4.0     # via wandb
docutils==0.15.2          # via botocore
dparse==0.5.1             # via safety
entrypoints==0.3          # via nbconvert
filelock==3.0.12          # via virtualenv
gitdb==4.0.5              # via gitpython
gitpython==3.1.7          # via bandit, wandb
gql==0.2.0                # via wandb
graphql-core==1.1         # via gql
identify==1.4.29          # via pre-commit
idna==2.10                # via -c requirements.txt, requests
iniconfig==1.0.1          # via pytest
ipykernel==5.3.4          # via notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.17.0           # via ipykernel
isort==4.3.21             # via pylint
itermplot==0.331          # via -r requirements-dev.in
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via jupyterlab, jupyterlab-server, nbconvert, notebook
jmespath==0.10.0          # via boto3, botocore
json5==0.9.5              # via jupyterlab-server
jsonschema==3.2.0         # via jupyterlab-server, nbformat
jupyter-client==6.1.6     # via ipykernel, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-server==1.2.0  # via jupyterlab
jupyterlab==2.2.4         # via -r requirements-dev.in
kiwisolver==1.2.0         # via -c requirements.txt, matplotlib
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.0         # via -c requirements.txt, -r requirements-dev.in, itermplot
mccabe==0.6.1             # via pylint
mistune==0.8.4            # via nbconvert
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.782               # via -r requirements-dev.in
nbconvert==5.6.1          # via notebook
nbformat==5.0.7           # via nbconvert, notebook
nodeenv==1.5.0            # via pre-commit
notebook==6.1.1           # via jupyterlab, jupyterlab-server
numpy==1.19.1             # via -c requirements.txt, itermplot, matplotlib
nvidia-ml-py3==7.352.0    # via -c requirements.txt, wandb
packaging==20.4           # via -c requirements.txt, bleach, dparse, pytest, safety
pandocfilters==1.4.2      # via nbconvert
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pathtools==0.1.2          # via watchdog
pbr==5.4.5                # via stevedore
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via -c requirements.txt, matplotlib
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements-dev.in
prometheus-client==0.8.0  # via notebook
promise==2.3              # via gql, graphql-core
prompt-toolkit==3.0.5     # via ipython
psutil==5.7.2             # via wandb
ptyprocess==0.6.0         # via pexpect, terminado
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via -r requirements-dev.in
pycparser==2.20           # via cffi
pydocstyle==5.0.2         # via -r requirements-dev.in
pygments==2.6.1           # via ipython, nbconvert
pylint==2.5.3             # via -r requirements-dev.in
pyparsing==2.4.7          # via -c requirements.txt, matplotlib, packaging
pyrsistent==0.16.0        # via jsonschema
pytest==6.0.1             # via -r requirements-dev.in
python-dateutil==2.8.1    # via -c requirements.txt, botocore, jupyter-client, matplotlib, wandb
pyyaml==5.3.1             # via -c requirements.txt, -r requirements-dev.in, bandit, dparse, pre-commit, wandb
pyzmq==19.0.2             # via jupyter-client, notebook
regex==2020.7.14          # via black
requests==2.24.0          # via -c requirements.txt, gql, jupyterlab-server, safety, wandb
s3transfer==0.3.3         # via boto3
safety==1.9.0             # via -r requirements-dev.in
send2trash==1.5.0         # via notebook
sentry-sdk==0.16.3        # via wandb
shortuuid==1.0.1          # via wandb
six==1.15.0               # via -c requirements.txt, argon2-cffi, astroid, bandit, bleach, cycler, docker-pycreds, gql, graphql-core, itermplot, jsonschema, packaging, promise, pyrsistent, python-dateutil, traitlets, virtualenv, wandb
smmap==3.0.4              # via gitdb
snowballstemmer==2.0.0    # via pydocstyle
stevedore==3.2.0          # via bandit
subprocess32==3.5.4       # via wandb
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.1              # via -c requirements.txt, black, dparse, pre-commit, pylint, pytest
torch-summary==1.4.1      # via -r requirements-dev.in
tornado==6.0.4            # via ipykernel, jupyter-client, jupyterlab, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
urllib3==1.25.10          # via -c requirements.txt, botocore, requests, sentry-sdk
virtualenv==20.0.31       # via pre-commit
wandb==0.9.4              # via -r requirements-dev.in
watchdog==0.10.3          # via wandb
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
wrapt==1.11.2             # via -c requirements.txt, astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
